--- eo-runtime/src/main/java/org/eolang/Oname.java
@@ -1,59 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2016-2022 Objectionary.com
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.eolang;
-
-/**
- * Object name, as in source code.
- *
- * @since 0.28.12
- */
-public final class Oname {
-
-    /**
-     * Object to get name.
-     */
-    private final Phi phi;
-
-    /**
-     * Ctor.
-     *
-     * @param phi The phi
-     */
-    Oname(final Phi phi) {
-        this.phi = phi;
-    }
-
-    @Override
-    public String toString() {
-        String txt = this.phi.getClass().getSimpleName();
-        final XmirObject xmir = this.getClass().getAnnotation(XmirObject.class);
-        if (null != xmir) {
-            txt = xmir.oname();
-            if ("@".equals(txt)) {
-                txt = "φ";
-            }
-        }
-        return txt;
-    }
-}
